fix: production hardening from code review
- #4: Support half-hour timezone offsets (float instead of int)
- #6: Add AbortController to cancel SSE stream on session switch
- #7: Fix TOOL_CALL_END report detection to match by toolCallId
- #8: Process remaining buffer after SSE stream ends
- #9: Clamp end_epoch to now in waf_bypass
- #10: Initialize result before poll loop in waf_patrol
- #11: Handle malformed JA4 cipher/ext counts gracefully
- #13: signOut sets user to false (login form) not null (loading)
- #14: Reset textarea height after submit

+++ agent.py
@@ -557,7 +557,7 @@ def create_app():
         tz_offset = forwarded.get("userTimezoneOffset")
         if tz_offset is not None:
             from tools.session_state import set_user_timezone
-            set_user_timezone(int(tz_offset))
+            set_user_timezone(float(tz_offset))
 
         msg_seq = int(time.time() * 1000)  # timestamp-based seq for DDB ordering
         return StreamingResponse(
